+++ esp-hal-common/src/fmt.rs
@@ -1,10 +1,16 @@
 #![macro_use]
 #![allow(unused_macros)]
 
+#[cfg(all(feature = "defmt", feature = "log"))]
+compile_error!("You may not enable both `defmt` and `log` features.");
+
 macro_rules! assert {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::assert!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::assert!($($x)*);
         }
     };
 }
@@ -12,7 +18,10 @@ macro_rules! assert {
 macro_rules! assert_eq {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::assert_eq!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::assert_eq!($($x)*);
         }
     };
 }
@@ -20,7 +29,10 @@ macro_rules! assert_eq {
 macro_rules! assert_ne {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::assert_ne!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::assert_ne!($($x)*);
         }
     };
 }
@@ -28,7 +40,10 @@ macro_rules! assert_ne {
 macro_rules! debug_assert {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::debug_assert!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::debug_assert!($($x)*);
         }
     };
 }
@@ -36,7 +51,10 @@ macro_rules! debug_assert {
 macro_rules! debug_assert_eq {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::debug_assert_eq!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::debug_assert_eq!($($x)*);
         }
     };
 }
@@ -44,7 +62,10 @@ macro_rules! debug_assert_eq {
 macro_rules! debug_assert_ne {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::debug_assert_ne!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::debug_assert_ne!($($x)*);
         }
     };
 }
@@ -52,7 +73,10 @@ macro_rules! debug_assert_ne {
 macro_rules! todo {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::todo!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::todo!($($x)*);
         }
     };
 }
@@ -60,7 +84,10 @@ macro_rules! todo {
 macro_rules! unreachable {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::unreachable!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::unreachable!($($x)*);
         }
     };
 }
@@ -68,7 +95,10 @@ macro_rules! unreachable {
 macro_rules! panic {
     ($($x:tt)*) => {
         {
+            #[cfg(not(feature = "defmt"))]
             ::core::panic!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::panic!($($x)*);
         }
     };
 }
@@ -78,7 +108,9 @@ macro_rules! trace {
         {
             #[cfg(feature = "log")]
             ::log::trace!($s $(, $x)*);
-            #[cfg(not(feature = "log"))]
+            #[cfg(feature = "defmt")]
+            ::defmt::trace!($s $(, $x)*);
+            #[cfg(not(any(feature = "log", feature="defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -89,7 +121,9 @@ macro_rules! debug {
         {
             #[cfg(feature = "log")]
             ::log::debug!($s $(, $x)*);
-            #[cfg(not(feature = "log"))]
+            #[cfg(feature = "defmt")]
+            ::defmt::debug!($s $(, $x)*);
+            #[cfg(not(any(feature = "log", feature="defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -100,7 +134,9 @@ macro_rules! info {
         {
             #[cfg(feature = "log")]
             ::log::info!($s $(, $x)*);
-            #[cfg(not(feature = "log"))]
+            #[cfg(feature = "defmt")]
+            ::defmt::info!($s $(, $x)*);
+            #[cfg(not(any(feature = "log", feature="defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -111,7 +147,9 @@ macro_rules! warn {
         {
             #[cfg(feature = "log")]
             ::log::warn!($s $(, $x)*);
-            #[cfg(not(feature = "log"))]
+            #[cfg(feature = "defmt")]
+            ::defmt::warn!($s $(, $x)*);
+            #[cfg(not(any(feature = "log", feature="defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -122,12 +160,22 @@ macro_rules! error {
         {
             #[cfg(feature = "log")]
             ::log::error!($s $(, $x)*);
-            #[cfg(not(feature = "log"))]
+            #[cfg(feature = "defmt")]
+            ::defmt::error!($s $(, $x)*);
+            #[cfg(not(any(feature = "log", feature="defmt")))]
             let _ = ($( & $x ),*);
         }
     };
 }
 
+#[cfg(feature = "defmt")]
+macro_rules! unwrap {
+    ($($x:tt)*) => {
+        ::defmt::unwrap!($($x)*)
+    };
+}
+
+#[cfg(not(feature = "defmt"))]
 macro_rules! unwrap {
     ($arg:expr) => {
         match $crate::fmt::Try::into_result($arg) {
@@ -144,5 +192,34 @@ macro_rules! unwrap {
                 ::core::panic!("unwrap of `{}` failed: {}: {:?}", ::core::stringify!($arg), ::core::format_args!($($msg,)*), e);
             }
         }
+    }
+}
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub struct NoneError;
+
+pub trait Try {
+    type Ok;
+    type Error;
+    fn into_result(self) -> Result<Self::Ok, Self::Error>;
+}
+
+impl<T> Try for Option<T> {
+    type Ok = T;
+    type Error = NoneError;
+
+    #[inline]
+    fn into_result(self) -> Result<T, NoneError> {
+        self.ok_or(NoneError)
+    }
+}
+
+impl<T, E> Try for Result<T, E> {
+    type Ok = T;
+    type Error = E;
+
+    #[inline]
+    fn into_result(self) -> Self {
+        self
     }
 }